+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "scratch-manager" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Unreleased
+
+- Add configuration for archive threshold, default to 90 days
+- Add configuration for an auto archive mode, default to off
+
 ## 0.1.0
 
 - Initial release